Moved tracker to mixins, created utils dir
